UIPCIR-5: Unsaved changes modal for Fast add form
I added state to the CreateRecordForm component
so that it could manage a modal window, and imported
the same confirmation modal stripesFinalForm uses for
navigation checks.

Note that we cannot use the NavigationCheck option
that StripesFinalForm offers here, because the form used
here does not change the URL. It would be really helpful if
stripesFinalForm included a way you could ask for the confirmation
modal to be triggered on form close in addition to triggering
on navigation change.

refs: https://issues.folio.org/browse/UIPCIR-5